pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soil-shazam-data-collector"
version = "0.1.0"
description = "Automated soil image dataset mining pipeline using CLIP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "open-clip-torch>=2.24.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "Pillow>=10.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "bing-image-downloader>=1.1.2",
    "imagededup>=0.3.2",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu124"

[project.scripts]
soil-collector = "soil_collector.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
